extract PackageTree to pkgtree package
fix tests
