+++ deduce_test.go
@@ -626,27 +626,3 @@ func ufmt(u *url.URL) string {
 	return fmt.Sprintf("host=%q, path=%q, opaque=%q, scheme=%q, user=%#v, pass=%#v, rawpath=%q, rawq=%q, frag=%q",
 		u.Host, u.Path, u.Opaque, u.Scheme, user, pass, u.RawPath, u.RawQuery, u.Fragment)
 }
-
-func TestIsStdLib(t *testing.T) {
-	fix := []struct {
-		ip string
-		is bool
-	}{
-		{"appengine", true},
-		{"net/http", true},
-		{"github.com/anything", false},
-		{"foo", true},
-	}
-
-	for _, f := range fix {
-		r := doIsStdLib(f.ip)
-		if r != f.is {
-			if r {
-				t.Errorf("%s was marked stdlib but should not have been", f.ip)
-			} else {
-				t.Errorf("%s was not marked stdlib but should have been", f.ip)
-
-			}
-		}
-	}
-}
