+++ pkgtree/pkgtree_test.go
@@ -1,4 +1,4 @@
-package gps
+package pkgtree
 
 import (
 	"fmt"
@@ -12,7 +12,37 @@ import (
 	"runtime"
 	"strings"
 	"testing"
+
+	"github.com/sdboyer/gps/fs"
 )
+
+func init() {
+	MockIsStdLib()
+}
+
+func TestIsStdLib(t *testing.T) {
+	fix := []struct {
+		ip string
+		is bool
+	}{
+		{"appengine", true},
+		{"net/http", true},
+		{"github.com/anything", false},
+		{"foo", true},
+	}
+
+	for _, f := range fix {
+		r := doIsStdLib(f.ip)
+		if r != f.is {
+			if r {
+				t.Errorf("%s was marked stdlib but should not have been", f.ip)
+			} else {
+				t.Errorf("%s was not marked stdlib but should have been", f.ip)
+
+			}
+		}
+	}
+}
 
 // PackageTree.ToReachMap() uses an easily separable algorithm, wmToReach(),
 // to turn a discovered set of packages and their imports into a proper pair of
@@ -31,7 +61,7 @@ func TestWorkmapToReach(t *testing.T) {
 	table := map[string]struct {
 		workmap  map[string]wm
 		rm       ReachMap
-		em       map[string]*ProblemImportError
+		em       map[string]*ErrProblemImport
 		backprop bool
 	}{
 		"single": {
@@ -129,8 +159,8 @@ func TestWorkmapToReach(t *testing.T) {
 					External: []string{"B/baz"},
 				},
 			},
-			em: map[string]*ProblemImportError{
-				"A": &ProblemImportError{
+			em: map[string]*ErrProblemImport{
+				"A": &ErrProblemImport{
 					ImportPath: "A",
 					Cause:      []string{"A/foo"},
 					Err:        missingPkgErr("A/foo"),
@@ -169,13 +199,13 @@ func TestWorkmapToReach(t *testing.T) {
 					External: []string{"B/baz"},
 				},
 			},
-			em: map[string]*ProblemImportError{
-				"A": &ProblemImportError{
+			em: map[string]*ErrProblemImport{
+				"A": &ErrProblemImport{
 					ImportPath: "A",
 					Cause:      []string{"A/foo", "A/bar"},
 					Err:        missingPkgErr("A/bar"),
 				},
-				"A/foo": &ProblemImportError{
+				"A/foo": &ErrProblemImport{
 					ImportPath: "A/foo",
 					Cause:      []string{"A/bar"},
 					Err:        missingPkgErr("A/bar"),
@@ -209,13 +239,13 @@ func TestWorkmapToReach(t *testing.T) {
 					External: []string{"B/baz"},
 				},
 			},
-			em: map[string]*ProblemImportError{
-				"A": &ProblemImportError{
+			em: map[string]*ErrProblemImport{
+				"A": &ErrProblemImport{
 					ImportPath: "A",
 					Cause:      []string{"A/foo"},
 					Err:        fmt.Errorf("err pkg"),
 				},
-				"A/foo": &ProblemImportError{
+				"A/foo": &ErrProblemImport{
 					ImportPath: "A/foo",
 					Err:        fmt.Errorf("err pkg"),
 				},
@@ -256,18 +286,18 @@ func TestWorkmapToReach(t *testing.T) {
 					External: []string{"B/baz"},
 				},
 			},
-			em: map[string]*ProblemImportError{
-				"A": &ProblemImportError{
+			em: map[string]*ErrProblemImport{
+				"A": &ErrProblemImport{
 					ImportPath: "A",
 					Cause:      []string{"A/foo", "A/bar"},
 					Err:        fmt.Errorf("err pkg"),
 				},
-				"A/foo": &ProblemImportError{
+				"A/foo": &ErrProblemImport{
 					ImportPath: "A/foo",
 					Cause:      []string{"A/bar"},
 					Err:        fmt.Errorf("err pkg"),
 				},
-				"A/bar": &ProblemImportError{
+				"A/bar": &ErrProblemImport{
 					ImportPath: "A/bar",
 					Err:        fmt.Errorf("err pkg"),
 				},
@@ -317,8 +347,8 @@ func TestWorkmapToReach(t *testing.T) {
 					External: []string{"B/baz"},
 				},
 			},
-			em: map[string]*ProblemImportError{
-				"A/bar": &ProblemImportError{
+			em: map[string]*ErrProblemImport{
+				"A/bar": &ErrProblemImport{
 					ImportPath: "A/bar",
 					Err:        fmt.Errorf("err pkg"),
 				},
@@ -436,7 +466,7 @@ func TestWorkmapToReach(t *testing.T) {
 		// needed
 		t.Run(name, func(t *testing.T) {
 			if fix.em == nil {
-				fix.em = make(map[string]*ProblemImportError)
+				fix.em = make(map[string]*ErrProblemImport)
 			}
 
 			rm, em := wmToReach(fix.workmap, fix.backprop)
@@ -453,7 +483,7 @@ func TestWorkmapToReach(t *testing.T) {
 }
 
 func TestListPackagesNoDir(t *testing.T) {
-	out, err := ListPackages(filepath.Join(getwd(t), "_testdata", "notexist"), "notexist")
+	out, err := ListPackages(filepath.Join(getTestdataRootDir(t), "notexist"), "notexist")
 	if err == nil {
 		t.Error("ListPackages should have errored on pointing to a nonexistent dir")
 	}
@@ -463,7 +493,7 @@ func TestListPackagesNoDir(t *testing.T) {
 }
 
 func TestListPackages(t *testing.T) {
-	srcdir := filepath.Join(getwd(t), "_testdata", "src")
+	srcdir := filepath.Join(getTestdataRootDir(t), "src")
 	j := func(s ...string) string {
 		return filepath.Join(srcdir, filepath.Join(s...))
 	}
@@ -1103,7 +1133,7 @@ func TestListPackages(t *testing.T) {
 						},
 					},
 					"relimport/dotdot": {
-						Err: &LocalImportsError{
+						Err: &ErrLocalImports{
 							Dir:        j("relimport/dotdot"),
 							ImportPath: "relimport/dotdot",
 							LocalImports: []string{
@@ -1112,7 +1142,7 @@ func TestListPackages(t *testing.T) {
 						},
 					},
 					"relimport/dotslash": {
-						Err: &LocalImportsError{
+						Err: &ErrLocalImports{
 							Dir:        j("relimport/dotslash"),
 							ImportPath: "relimport/dotslash",
 							LocalImports: []string{
@@ -1121,7 +1151,7 @@ func TestListPackages(t *testing.T) {
 						},
 					},
 					"relimport/dotdotslash": {
-						Err: &LocalImportsError{
+						Err: &ErrLocalImports{
 							Dir:        j("relimport/dotdotslash"),
 							ImportPath: "relimport/dotdotslash",
 							LocalImports: []string{
@@ -1316,9 +1346,9 @@ func TestListPackagesNoPerms(t *testing.T) {
 	}
 	defer os.RemoveAll(tmp)
 
-	srcdir := filepath.Join(getwd(t), "_testdata", "src", "ren")
+	srcdir := filepath.Join(getTestdataRootDir(t), "src", "ren")
 	workdir := filepath.Join(tmp, "ren")
-	copyDir(srcdir, workdir)
+	fs.CopyDir(srcdir, workdir)
 
 	// chmod the simple dir and m1p/b.go file so they can't be read
 	err = os.Chmod(filepath.Join(workdir, "simple"), 0)
@@ -1385,186 +1415,9 @@ func TestListPackagesNoPerms(t *testing.T) {
 	}
 }
 
-func TestFlattenReachMap(t *testing.T) {
-	// There's enough in the 'varied' test case to test most of what matters
-	vptree, err := ListPackages(filepath.Join(getwd(t), "_testdata", "src", "github.com", "example", "varied"), "github.com/example/varied")
-	if err != nil {
-		t.Fatalf("listPackages failed on varied test case: %s", err)
-	}
-
-	var expect []string
-	var name string
-	var ignore map[string]bool
-	var stdlib, main, tests bool
-
-	validate := func() {
-		rm, em := vptree.ToReachMap(main, tests, true, ignore)
-		if len(em) != 0 {
-			t.Errorf("Should not have any error pkgs from ToReachMap, got %s", em)
-		}
-		result := rm.Flatten(stdlib)
-		if !reflect.DeepEqual(expect, result) {
-			t.Errorf("Wrong imports in %q case:\n\t(GOT): %s\n\t(WNT): %s", name, result, expect)
-		}
-	}
-
-	all := []string{
-		"encoding/binary",
-		"github.com/Masterminds/semver",
-		"github.com/sdboyer/gps",
-		"go/parser",
-		"hash",
-		"net/http",
-		"os",
-		"sort",
-	}
-
-	// helper to rewrite expect, except for a couple packages
-	//
-	// this makes it easier to see what we're taking out on each test
-	except := func(not ...string) {
-		expect = make([]string, len(all)-len(not))
-
-		drop := make(map[string]bool)
-		for _, npath := range not {
-			drop[npath] = true
-		}
-
-		k := 0
-		for _, path := range all {
-			if !drop[path] {
-				expect[k] = path
-				k++
-			}
-		}
-	}
-
-	// everything on
-	name = "simple"
-	except()
-	stdlib, main, tests = true, true, true
-	validate()
-
-	// turning off stdlib should cut most things, but we need to override the
-	// function
-	isStdLib = doIsStdLib
-	name = "no stdlib"
-	stdlib = false
-	except("encoding/binary", "go/parser", "hash", "net/http", "os", "sort")
-	validate()
-	// Restore stdlib func override
-	overrideIsStdLib()
-
-	// stdlib back in; now exclude tests, which should just cut one
-	name = "no tests"
-	stdlib, tests = true, false
-	except("encoding/binary")
-	validate()
-
-	// Now skip main, which still just cuts out one
-	name = "no main"
-	main, tests = false, true
-	except("net/http")
-	validate()
-
-	// No test and no main, which should be additive
-	name = "no test, no main"
-	main, tests = false, false
-	except("net/http", "encoding/binary")
-	validate()
-
-	// now, the ignore tests. turn main and tests back on
-	main, tests = true, true
-
-	// start with non-matching
-	name = "non-matching ignore"
-	ignore = map[string]bool{
-		"nomatch": true,
-	}
-	except()
-	validate()
-
-	// should have the same effect as ignoring main
-	name = "ignore the root"
-	ignore = map[string]bool{
-		"github.com/example/varied": true,
-	}
-	except("net/http")
-	validate()
-
-	// now drop a more interesting one
-	name = "ignore simple"
-	ignore = map[string]bool{
-		"github.com/example/varied/simple": true,
-	}
-	// we get github.com/sdboyer/gps from m1p, too, so it should still be there
-	except("go/parser")
-	validate()
-
-	// now drop two
-	name = "ignore simple and namemismatch"
-	ignore = map[string]bool{
-		"github.com/example/varied/simple":       true,
-		"github.com/example/varied/namemismatch": true,
-	}
-	except("go/parser", "github.com/Masterminds/semver")
-	validate()
-
-	// make sure tests and main play nice with ignore
-	name = "ignore simple and namemismatch, and no tests"
-	tests = false
-	except("go/parser", "github.com/Masterminds/semver", "encoding/binary")
-	validate()
-	name = "ignore simple and namemismatch, and no main"
-	main, tests = false, true
-	except("go/parser", "github.com/Masterminds/semver", "net/http")
-	validate()
-	name = "ignore simple and namemismatch, and no main or tests"
-	main, tests = false, false
-	except("go/parser", "github.com/Masterminds/semver", "net/http", "encoding/binary")
-	validate()
-
-	main, tests = true, true
-
-	// ignore two that should knock out gps
-	name = "ignore both importers"
-	ignore = map[string]bool{
-		"github.com/example/varied/simple": true,
-		"github.com/example/varied/m1p":    true,
-	}
-	except("sort", "github.com/sdboyer/gps", "go/parser")
-	validate()
-
-	// finally, directly ignore some external packages
-	name = "ignore external"
-	ignore = map[string]bool{
-		"github.com/sdboyer/gps": true,
-		"go/parser":              true,
-		"sort":                   true,
-	}
-	except("sort", "github.com/sdboyer/gps", "go/parser")
-	validate()
-
-	// The only thing varied *doesn't* cover is disallowed path patterns
-	ptree, err := ListPackages(filepath.Join(getwd(t), "_testdata", "src", "disallow"), "disallow")
-	if err != nil {
-		t.Fatalf("ListPackages failed on disallow test case: %s", err)
-	}
-
-	rm, em := ptree.ToReachMap(false, false, true, nil)
-	if len(em) != 0 {
-		t.Errorf("Should not have any error packages from ToReachMap, got %s", em)
-	}
-	result := rm.Flatten(true)
-	expect = []string{"github.com/sdboyer/gps", "hash", "sort"}
-	if !reflect.DeepEqual(expect, result) {
-		t.Errorf("Wrong imports in %q case:\n\t(GOT): %s\n\t(WNT): %s", name, result, expect)
-	}
-}
-
 func TestToReachMap(t *testing.T) {
 	// There's enough in the 'varied' test case to test most of what matters
-	vptree, err := ListPackages(filepath.Join(getwd(t), "_testdata", "src", "github.com", "example", "varied"), "github.com/example/varied")
+	vptree, err := ListPackages(filepath.Join(getTestdataRootDir(t), "src", "github.com", "example", "varied"), "github.com/example/varied")
 	if err != nil {
 		t.Fatalf("ListPackages failed on varied test case: %s", err)
 	}
@@ -1826,9 +1679,186 @@ func TestToReachMap(t *testing.T) {
 	validate()
 }
 
+func TestFlattenReachMap(t *testing.T) {
+	// There's enough in the 'varied' test case to test most of what matters
+	vptree, err := ListPackages(filepath.Join(getTestdataRootDir(t), "src", "github.com", "example", "varied"), "github.com/example/varied")
+	if err != nil {
+		t.Fatalf("listPackages failed on varied test case: %s", err)
+	}
+
+	var expect []string
+	var name string
+	var ignore map[string]bool
+	var stdlib, main, tests bool
+
+	validate := func() {
+		rm, em := vptree.ToReachMap(main, tests, true, ignore)
+		if len(em) != 0 {
+			t.Errorf("Should not have any error pkgs from ToReachMap, got %s", em)
+		}
+		result := rm.Flatten(stdlib)
+		if !reflect.DeepEqual(expect, result) {
+			t.Errorf("Wrong imports in %q case:\n\t(GOT): %s\n\t(WNT): %s", name, result, expect)
+		}
+	}
+
+	all := []string{
+		"encoding/binary",
+		"github.com/Masterminds/semver",
+		"github.com/sdboyer/gps",
+		"go/parser",
+		"hash",
+		"net/http",
+		"os",
+		"sort",
+	}
+
+	// helper to rewrite expect, except for a couple packages
+	//
+	// this makes it easier to see what we're taking out on each test
+	except := func(not ...string) {
+		expect = make([]string, len(all)-len(not))
+
+		drop := make(map[string]bool)
+		for _, npath := range not {
+			drop[npath] = true
+		}
+
+		k := 0
+		for _, path := range all {
+			if !drop[path] {
+				expect[k] = path
+				k++
+			}
+		}
+	}
+
+	// everything on
+	name = "simple"
+	except()
+	stdlib, main, tests = true, true, true
+	validate()
+
+	// turning off stdlib should cut most things, but we need to override the
+	// function
+	isStdLib = doIsStdLib
+	name = "no stdlib"
+	stdlib = false
+	except("encoding/binary", "go/parser", "hash", "net/http", "os", "sort")
+	validate()
+	// restore stdlib func override
+	MockIsStdLib()
+
+	// stdlib back in; now exclude tests, which should just cut one
+	name = "no tests"
+	stdlib, tests = true, false
+	except("encoding/binary")
+	validate()
+
+	// Now skip main, which still just cuts out one
+	name = "no main"
+	main, tests = false, true
+	except("net/http")
+	validate()
+
+	// No test and no main, which should be additive
+	name = "no test, no main"
+	main, tests = false, false
+	except("net/http", "encoding/binary")
+	validate()
+
+	// now, the ignore tests. turn main and tests back on
+	main, tests = true, true
+
+	// start with non-matching
+	name = "non-matching ignore"
+	ignore = map[string]bool{
+		"nomatch": true,
+	}
+	except()
+	validate()
+
+	// should have the same effect as ignoring main
+	name = "ignore the root"
+	ignore = map[string]bool{
+		"github.com/example/varied": true,
+	}
+	except("net/http")
+	validate()
+
+	// now drop a more interesting one
+	name = "ignore simple"
+	ignore = map[string]bool{
+		"github.com/example/varied/simple": true,
+	}
+	// we get github.com/sdboyer/gps from m1p, too, so it should still be there
+	except("go/parser")
+	validate()
+
+	// now drop two
+	name = "ignore simple and namemismatch"
+	ignore = map[string]bool{
+		"github.com/example/varied/simple":       true,
+		"github.com/example/varied/namemismatch": true,
+	}
+	except("go/parser", "github.com/Masterminds/semver")
+	validate()
+
+	// make sure tests and main play nice with ignore
+	name = "ignore simple and namemismatch, and no tests"
+	tests = false
+	except("go/parser", "github.com/Masterminds/semver", "encoding/binary")
+	validate()
+	name = "ignore simple and namemismatch, and no main"
+	main, tests = false, true
+	except("go/parser", "github.com/Masterminds/semver", "net/http")
+	validate()
+	name = "ignore simple and namemismatch, and no main or tests"
+	main, tests = false, false
+	except("go/parser", "github.com/Masterminds/semver", "net/http", "encoding/binary")
+	validate()
+
+	main, tests = true, true
+
+	// ignore two that should knock out gps
+	name = "ignore both importers"
+	ignore = map[string]bool{
+		"github.com/example/varied/simple": true,
+		"github.com/example/varied/m1p":    true,
+	}
+	except("sort", "github.com/sdboyer/gps", "go/parser")
+	validate()
+
+	// finally, directly ignore some external packages
+	name = "ignore external"
+	ignore = map[string]bool{
+		"github.com/sdboyer/gps": true,
+		"go/parser":              true,
+		"sort":                   true,
+	}
+	except("sort", "github.com/sdboyer/gps", "go/parser")
+	validate()
+
+	// The only thing varied *doesn't* cover is disallowed path patterns
+	ptree, err := ListPackages(filepath.Join(getTestdataRootDir(t), "src", "disallow"), "disallow")
+	if err != nil {
+		t.Fatalf("ListPackages failed on disallow test case: %s", err)
+	}
+
+	rm, em := ptree.ToReachMap(false, false, true, nil)
+	if len(em) != 0 {
+		t.Errorf("Should not have any error packages from ToReachMap, got %s", em)
+	}
+	result := rm.Flatten(true)
+	expect = []string{"github.com/sdboyer/gps", "hash", "sort"}
+	if !reflect.DeepEqual(expect, result) {
+		t.Errorf("Wrong imports in %q case:\n\t(GOT): %s\n\t(WNT): %s", name, result, expect)
+	}
+}
+
 // Verify that we handle import cycles correctly - drop em all
 func TestToReachMapCycle(t *testing.T) {
-	ptree, err := ListPackages(filepath.Join(getwd(t), "_testdata", "src", "cycle"), "cycle")
+	ptree, err := ListPackages(filepath.Join(getTestdataRootDir(t), "src", "cycle"), "cycle")
 	if err != nil {
 		t.Fatalf("ListPackages failed on cycle test case: %s", err)
 	}
@@ -1848,10 +1878,10 @@ func TestToReachMapCycle(t *testing.T) {
 	}
 }
 
-func getwd(t *testing.T) string {
+func getTestdataRootDir(t *testing.T) string {
 	cwd, err := os.Getwd()
 	if err != nil {
 		t.Fatal(err)
 	}
-	return cwd
+	return filepath.Join(cwd, "..", "_testdata")
 }
